+++ app/components/UI/AssetOverview/TokenDetails/TokenDetails.test.tsx
@@ -16,11 +16,11 @@ import {
 } from '../../../../selectors/networkController';
 // eslint-disable-next-line import/no-namespace
 import * as reactRedux from 'react-redux';
-// eslint-disable-next-line import/no-namespace
-import * as StakeConstants from '../../Stake/constants';
 import { strings } from '../../../../../locales/i18n';
 import { selectMultichainAssetsRates } from '../../../../selectors/multichain';
 import { selectIsEvmNetworkSelected } from '../../../../selectors/multichainNetworkController';
+import { selectStablecoinLendingEnabledFlag } from '../../Earn/selectors/featureFlags';
+
 jest.mock('../../../../core/Engine', () => ({
   getTotalEvmFiatAccountBalance: jest.fn(),
   context: {
@@ -133,16 +133,9 @@ const mockTokenMarketDataByChainId: Record<
   },
 };
 
-let isStablecoinLendingFeatureEnabledSpy: jest.SpyInstance;
-
 describe('TokenDetails', () => {
   beforeAll(() => {
     jest.resetAllMocks();
-  });
-  beforeEach(() => {
-    isStablecoinLendingFeatureEnabledSpy = jest
-      .spyOn(StakeConstants, 'isStablecoinLendingFeatureEnabled')
-      .mockReturnValue(false);
   });
 
   const mockSelectors = () => {
@@ -194,6 +187,8 @@ describe('TokenDetails', () => {
           return {};
         case selectIsEvmNetworkSelected:
           return true;
+        case selectStablecoinLendingEnabledFlag:
+          return false;
         default:
           return undefined;
       }
@@ -335,9 +330,45 @@ describe('TokenDetails', () => {
   });
 
   it('renders EarnEmptyStateCta if asset can be lent and balance is not zero', () => {
-    isStablecoinLendingFeatureEnabledSpy.mockReturnValue(true);
+    const useSelectorSpy = jest.spyOn(reactRedux, 'useSelector');
+    const SELECTOR_MOCKS = {
+      selectTokenMarketDataByChainId: {},
+      selectConversionRateBySymbol: mockExchangeRate,
+      selectNativeCurrencyByChainId: 'ETH',
+    } as const;
 
-    mockSelectors();
+    useSelectorSpy.mockImplementation((selectorOrCallback) => {
+      if (typeof selectorOrCallback === 'function') {
+        const selectorString = selectorOrCallback.toString();
+        const matchedSelector = Object.keys(SELECTOR_MOCKS).find((key) =>
+          selectorString.includes(key),
+        );
+        if (matchedSelector) {
+          return SELECTOR_MOCKS[matchedSelector as keyof typeof SELECTOR_MOCKS];
+        }
+      }
+
+      switch (selectorOrCallback) {
+        case selectTokenList:
+          return mockAssets;
+        case selectContractExchangeRates:
+          return {};
+        case selectConversionRate:
+          return mockExchangeRate;
+        case selectCurrentCurrency:
+          return mockCurrentCurrency;
+        case selectProviderConfig:
+          return { ticker: 'ETH' };
+        case selectEvmTicker:
+          return 'ETH';
+        case selectMultichainAssetsRates:
+          return {};
+        case selectStablecoinLendingEnabledFlag:
+          return true;
+        default:
+          return undefined;
+      }
+    });
 
     const { getByText } = renderWithProvider(<TokenDetails asset={mockDAI} />, {
       state: initialState,
@@ -356,8 +387,6 @@ describe('TokenDetails', () => {
   });
 
   it('does not render EarnEmptyCta if asset can be lent but balance is zero', () => {
-    isStablecoinLendingFeatureEnabledSpy.mockReturnValue(true);
-
     mockSelectors();
 
     const mockDaiWithoutBalance = {